backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.8
pinecone-client==3.0.2
python-multipart==0.0.9
pydantic==2.6.1
httpx==0.26.0 